Explicit casts to avoid conversion warnings: std::size_t
Visual Studio warns about signed/unsigned conversion and limited range of types.

+++ src/solvers/sat/satcheck_minisat2.cpp
@@ -53,9 +53,9 @@ tvt satcheck_minisat2_baset<T>::l_get(literalt a) const
 
   using Minisat::lbool;
 
-  if(solver->model[a.var_no()]==l_True)
+  if(solver->model[narrow_cast<std::size_t>(a.var_no())] == l_True)
     result=tvt(true);
-  else if(solver->model[a.var_no()]==l_False)
+  else if(solver->model[narrow_cast<std::size_t>(a.var_no())] == l_False)
     result=tvt(false);
   else
     return tvt::unknown();
